+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+* 0.1.17
+    - 修复了 AJAX 请求不发送 NMP 指定 Header 的问题
+    - 使得 `io.hooks.*` 返回值用来修改传入参数
+    - 修复了 `navigator` 匹配路径为字符串的问题
+    - `navigator` 现在支持二级菜单及设置显示隐藏
+    - 升级了 `serverIO` 模块，以适配新版本 NMP 接口并更好地显示错误信息
+    - 修复了 `extension/underscore` 重的一些 bug
+    - 修复了 `mvc/BaseView` 中 `popDialog` 的一些问题
+    - 给表单页面增加了多个异步确认扩展点
+    - 表单页在验证出错后现在会默认将页面滚动到第一个出错的位置
+
 * 0.1.16
     - 增加了可以对单个请求发送器设定不显示 loading 的功能
     - 对 `underscore` 的依赖升级到 `1.6.0`